requirements.txt
attrs==21.4.0
httpx==0.22.0
aiogram==2.19.0
sentry-sdk==1.13.0
beautifulsoup4==4.11.1
python-decouple==3.7
pyinstaller==5.7.0
requests==2.28.2